Break out indicator plant into separate component
